% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{capitalise_tags_and_keywords}
\alias{capitalise_tags_and_keywords}
\title{Capitalise tags and certain keywords}
\usage{
capitalise_tags_and_keywords(gedcom)
}
\arguments{
\item{gedcom}{A tidyged object.}
}
\value{
A tidyged object with appropriately capitalised tags and keywords.
}
\description{
This function capitalises all tags and certain values such as SEX values and DATE values.
}
\details{
The function also ensures certain values are lowercase such as PEDI
and ADOP values, and removes explicit GREGORIAN date escape sequences (as they are implied).
}
