% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{combine_spouse_age_lines}
\alias{combine_spouse_age_lines}
\title{Combine husband/wife age rows into one row}
\usage{
combine_spouse_age_lines(gedcom)
}
\arguments{
\item{gedcom}{A tidyged object.}
}
\value{
A tidyged object with husband/wife and age rows combined accordingly.
}
\description{
This function combines the HUSB and WIFE rows with their subordinate AGE rows in order
to make querying easier.
}
