% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{create_addr_records}
\alias{create_addr_records}
\title{Create custom address records}
\usage{
create_addr_records(gedcom)
}
\arguments{
\item{gedcom}{A tidyged object.}
}
\value{
An updated tidyged object with the addition of address records.
}
\description{
This function reduces duplication by creating address records and referencing
them in the relevant places in the object.
}
