% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{create_conc_lines}
\alias{create_conc_lines}
\title{Create CONCatenation lines}
\usage{
create_conc_lines(lines, char_limit)
}
\arguments{
\item{lines}{Lines of a tidyged object.}

\item{char_limit}{Character limit of line values.}
}
\value{
The same lines of the tidyged object, potentially with additional concatenation lines.
}
\description{
Create CONCatenation lines
}
