% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{create_cont_lines}
\alias{create_cont_lines}
\title{Create CONTinuation lines}
\usage{
create_cont_lines(lines)
}
\arguments{
\item{lines}{Lines of a tidyged object.}
}
\value{
The same lines of the tidyged object, potentially with additional continuation lines.
}
\description{
Create CONTinuation lines
}
