% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{remove_addr_records}
\alias{remove_addr_records}
\title{Remove custom address records}
\usage{
remove_addr_records(gedcom)
}
\arguments{
\item{gedcom}{A tidyged object.}
}
\value{
An updated tidyged object with address records moved to the appropriate locations.
}
\description{
This function removes the custom address records from a tidyged object and places
them in the relevant places in the object.
}
