% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{set_class_to_tidyged}
\alias{set_class_to_tidyged}
\title{Make a dataframe a tidyged object}
\usage{
set_class_to_tidyged(gedcom)
}
\arguments{
\item{gedcom}{A tibble with content consistent with that of a tidyged object.}
}
\value{
A tidyged object.
}
\description{
This function sets the attribute on a dataframe identifying it as a tidyged object.
}
