% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{update_header_with_filename}
\alias{update_header_with_filename}
\title{Update GEDCOM header with filename}
\usage{
update_header_with_filename(gedcom, filename)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{filename}{The name of the file (with extension).}
}
\value{
An updated tidyged object with the updated filename.
}
\description{
Update GEDCOM header with filename
}
\details{
If the file does not already have a filename, then one is not added.
}
